+++ HM/Tests/Files/NamespaceDirectoryNameUnitTest/mu-plugins/headered/loader.php
@@ -1,8 +1,0 @@
-<?php
-/**
- * Plugin Name: Headered Nested Plugin
- *
- * Entry point identified by its plugin header rather than its file name.
- */
-
-namespace HM\Coding\Standards\Headered;