+++ CHANGELOG.md
@@ -5,7 +5,7 @@ and fill in the contents as you go. This simplifies later release management. --
 
 ## 2.3.0
 
-- Allow a plugin's entry-point file -- the main file in a plugin folder under `plugins/`, `mu-plugins/`, or `client-mu-plugins/`, named `<dir>/<dir>.php` or `<dir>/plugin.php` or declaring a `Plugin Name` header -- to declare a namespace without being named `namespace.php` #XXX
+- Allow a plugin's entry-point file -- a single-file mu-plugin, or the main file in a plugin folder under `plugins/`, `mu-plugins/`, or `client-mu-plugins/` named `<dir>/<dir>.php` or `<dir>/plugin.php` -- to declare a namespace without being named `namespace.php` #XXX
 
 ## 2.2.1